+++ src/app/examples/http-example.component.ts
@@ -6,11 +6,23 @@ import { ibCrudToast } from '../inobeta-ui/http/http/messages.decorator';
 @Component({
   selector: 'app-test',
   template: `
-
-  <button (click)="decoratorTest()">Test decorator (success)</button>
-  <button (click)="decoratorTest('some-error')">Test decorator (error)</button>
-  <button (click)="otherTest()">Test 401</button>
-  <button (click)="noSpinner()">No Spinner test</button>
+<mat-grid-list [cols]="4"  class="ib-material-form-grid" rowHeight="60px">
+  <mat-grid-tile [colspan]="1" [rowspan]="1">
+    <button mat-raised-button color="primary" (click)="decoratorTest()">Test decorator (success)</button>
+  </mat-grid-tile>
+  <mat-grid-tile [colspan]="1" [rowspan]="1">
+    <button mat-raised-button color="primary" (click)="decoratorTest('some-error')">Test decorator (error)</button>
+  </mat-grid-tile>
+  <mat-grid-tile [colspan]="1" [rowspan]="1">
+    <button mat-raised-button color="primary" (click)="otherTest()">Test 401</button>
+  </mat-grid-tile>
+  <mat-grid-tile [colspan]="1" [rowspan]="1">
+    <button mat-raised-button color="primary" (click)="noSpinner()">No Spinner test</button>
+  </mat-grid-tile>
+  <mat-grid-tile [colspan]="1" [rowspan]="1">
+    <button mat-raised-button color="primary" (click)="getBlobData()">Test responseType 'blob'</button>
+  </mat-grid-tile>
+</mat-grid-list>
 
   <pre>
     {{ loadedData | json }}
@@ -32,8 +44,8 @@ export class HttpExampleComponent implements OnInit {
   }
 
   @ibCrudToast()
-  serviceCall(wrong = '') {
-    return this.h.get(`assets/i18n/it.json${wrong}`).pipe(
+  serviceCall(wrong = '', data = null) {
+    return this.h.get(`assets/i18n/it.json${wrong}`, data).pipe(
       map((x) => {
         console.log('additional map', x);
         return x;
@@ -46,6 +58,12 @@ export class HttpExampleComponent implements OnInit {
       console.log('final data', data);
     }, err => {
       console.log('error data', err);
+    });
+  }
+
+  getBlobData() {
+    return this.h.get('assets/i18n/it.json', {provaTest: 'bah'}, { responseType: 'blob'}).subscribe(data => {
+      console.log('blob data', data);
     });
   }