+++ src/app/inobeta-ui/ui/table/store/selectors/table.selectors.spec.ts
@@ -1,12 +1,9 @@
 import * as fromTable from '../reducers/table.reducer';
-import { selectTableState } from './table.selectors';
+import { ibTableSelectTotalRow } from './table.selectors';
 
 describe('Table Selectors', () => {
-  it('should select the feature state', () => {
-    const result = selectTableState({
-      [fromTable.tableFeatureKey]: {}
-    });
-
-    expect(result).toEqual({});
+  it('should select the totals feature state', () => {
+    const result = ibTableSelectTotalRow.projector(fromTable.initialState);
+    expect(result).toEqual(fromTable.initialState.totals);
   });
 });